+++ R/reexports.R
@@ -22,3 +22,7 @@ generics::glance
 #' @importFrom generics augment
 #' @export
 generics::augment
+
+#' @importFrom generics required_pkgs
+#' @export
+generics::required_pkgs