+++ NEWS.md
@@ -1,5 +1,7 @@
 # parsnip (development version)
 
+* `generics::required_pkgs()` was extended for `parsnip` objects. 
+
 * The `liquidSVM` engine for `svm_rbf()` was deprecated due to that package's removal from CRAN. (#425)
 
 * A new linear SVM model `svm_linear()` is now available with the `LiblineaR` engine (#424) and the `kernlab` engine (#438), and the `LiblineaR` engine is available for `logistic_reg()` as well (#429).